for bug http://en.wikipedia.org/wiki/Wikipedia_talk:AutoWikiBrowser/Bugs#Disc_full add error handling to look for System.IO.IOExeption in ErrorHandler and raise specific message; in UserSettings SavePrefs don't try copying settings if IOException has been raised
git-svn-id: https://svn.code.sf.net/p/autowikibrowser/code/AWB@3800 5227b50d-9349-4574-b055-4b0a7525e11c

+++ AWB/UserSettings.cs
@@ -234,7 +234,9 @@ namespace AutoWikiBrowser
             catch (Exception ex)
             {
                 ErrorHandler.Handle(ex);
-                if (File.Exists(SettingsFile + ".old"))
+
+                // don't attempt to write to disk if the error was IOException (disk full etc.)
+                if (!(ex is System.IO.IOException) && File.Exists(SettingsFile + ".old"))
                     File.Copy(SettingsFile + ".old", SettingsFile, true);
             }
         }
